requirements.txt
coverage==7.7.1
Flask==3.1.0
pytest==8.1.1
pytest-cov==6.0.0
gunicorn
finnhub-python
requests